requirements.txt
Pillow~=9.0.1
cryptography~=3.4.8
bleak~=0.21.1